BUG/MEDIUM: backend: Update hash to use unsigned int throughout
When we were generating a hash, it was done using an unsigned long.  When the hash was used
to select a backend, it was sent as an unsigned int.  This made it difficult to predict which
backend would be selected.

This patch updates get_hash, and the hash methods to use an unsigned int, to remain consistent
throughout the codebase.

This fix should be backported to 1.5 and probably in part to 1.4.
(cherry picked from commit bd57a9f977f60fcf7818f462953da3740e3bd010)
(cherry picked from commit 82456753cae9200aa1f4031f64c22c08e130f072)

+++ src/backend.c
@@ -138,7 +138,7 @@ struct server *get_server_sh(struct proxy *px, const char *addr, int len)
  */
 struct server *get_server_uh(struct proxy *px, char *uri, int uri_len)
 {
-	unsigned long hash = 0;
+	unsigned int hash = 0;
 	int c;
 	int slashes = 0;
 
@@ -182,7 +182,7 @@ struct server *get_server_uh(struct proxy *px, char *uri, int uri_len)
  */
 struct server *get_server_ph(struct proxy *px, const char *uri, int uri_len)
 {
-	unsigned long hash = 0;
+	unsigned int hash = 0;
 	const char *p;
 	const char *params;
 	int plen;
@@ -238,7 +238,7 @@ struct server *get_server_ph(struct proxy *px, const char *uri, int uri_len)
  */
 struct server *get_server_ph_post(struct session *s)
 {
-	unsigned long    hash = 0;
+	unsigned int     hash = 0;
 	struct http_txn *txn  = &s->txn;
 	struct buffer   *req  = s->req;
 	struct http_msg *msg  = &txn->req;
@@ -313,7 +313,7 @@ struct server *get_server_ph_post(struct session *s)
  */
 struct server *get_server_hh(struct session *s)
 {
-	unsigned long    hash = 0;
+	unsigned int     hash = 0;
 	struct http_txn *txn  = &s->txn;
 	struct http_msg *msg  = &txn->req;
 	struct proxy    *px   = s->be;
@@ -382,7 +382,7 @@ struct server *get_server_hh(struct session *s)
 
 struct server *get_server_rch(struct session *s)
 {
-	unsigned long    hash = 0;
+	unsigned int     hash = 0;
 	struct proxy    *px   = s->be;
 	unsigned long    len;
 	const char      *p;
